Split up the inferers as they have grown too long.
