+++ src/llmbo/__init__.py
@@ -1,14 +1,12 @@
 from .adapters import AnthropicAdapter
-from .llmbo import (
-    BatchInferer,
-    StructuredBatchInferer,
-)
+from .batch_inferer import BatchInferer
 from .models import (
     Manifest,
     ModelInput,
     ToolChoice,
 )
 from .registry import ModelAdapterRegistry
+from .structured_batch_inferer import StructuredBatchInferer
 
 # Register the model adapters
 ModelAdapterRegistry.register(r"(anthropic|claude)", AnthropicAdapter)
